refactor(output): split activity tests — handles (85) + dispatch (32)
